requirements.txt
# EduChat Requirements
# Install with: pip install -r requirements.txt

# FastAPI & Server
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# LLM & AI
google-genai==1.67.0

# LangChain (for RAG and Chat Memory)
langchain==0.3.13
langchain-core==0.3.63
langchain-google-genai>=0.2.0
langchain-community==0.3.13
langchain-text-splitters==0.3.8
langchain-huggingface>=0.1.0
langchain-groq>=0.2.1

# Document Processing
pypdf==5.1.0
python-docx==1.2.0

# Vector Store (for RAG embeddings)
faiss-cpu==1.9.0
numpy==1.26.4

# Database (Supabase/PostgreSQL)
psycopg2-binary==2.9.9
supabase==2.28.3